Remove predefined-colors
The script initially had a list of predefined colors to choose from.
This list is arbitrary and doesn't include the colorschemes that the
user has. This commit makes this list empty so that :NextColor will
populate it when it's invoked. Note that while we could populate this
list, it would affect the startup time of Vim, which is why we delegate
it to a future :NextColor call.